Added new firrtl example

+++ tests/utils/cmp_arc_btor.sh
@@ -5,7 +5,7 @@
 
 # Check that at least one argument was give
 if [[ $# -lt 1 ]]; then
-    echo "Usage: ./cmp_arc_btor.sh <filename>.fir" >&2
+    echo "Usage: ./cmp_arc_btor.sh <filename>.fir [optional: <outname>]" >&2
     exit 2
 fi
 
@@ -23,6 +23,14 @@ firtool --ir-fir $1 >> tmp_fir.mlir
 
 # Step 2: Convert mlir fir to hw
 circt-opt --lower-firrtl-to-hw tmp_fir.mlir >> tmp_hw.mlir
+echo " "
+echo "===================="
+echo "Output from hw:"       
+echo "===================="
+echo " "
+
+cat tmp_hw.mlir
+
 
 # Step 3: Convert hw to arc
 echo " "
@@ -30,22 +38,34 @@ echo "===================="
 echo "Output from arcilator:"
 echo "===================="
 echo " "
-arcilator tmp_hw.mlir
+arcilator --emit-mlir tmp_hw.mlir
 
 # Clean-up tmp files
 rm tmp_fir.mlir tmp_hw.mlir
 
 # Part 2: Run firrtl
-firrtl -E btor2 -i $1 -o tmp.btor2
-
-# Print out result
 echo " "
 echo "===================="
 echo "Output from firrtl formal:"
 echo "===================="
 echo " "
-cat tmp.btor2
+# Check that at least one argument was give
+if [[ $# -gt 1 ]]; then
+    firrtl --compiler sverilog -E btor2 -i $1 -o $2.btor2
 
-# Cleanup tmp files
-rm tmp.btor2
+    # Print out result
+    cat $2.btor2
+else
+    firrtl --compiler sverilog -E btor2 -i $1 -o tmp.btor2
+
+    # Print out result
+    cat tmp.btor2
+
+    # Cleanup tmp files
+    rm tmp.btor2
+fi
+
+
+
+
